+++ src/App.vue
@@ -4,8 +4,8 @@ import TheWelcome from './components/TheWelcome.vue'
 // import AttributePage from './components/AttributePage.vue'
 import SimpleComponentA from './components/SimpleComponentA.vue'
 import SimpleComponentB from './components/SimpleComponentB.vue'
-import AttributeComponent from './components/AttributeComponent.vue'
-import TalentComponent from './components/TalentComponent.vue'
+import AttributeComponent from './components/AttributesView.vue'
+import TalentComponent from './components/TalentsView.vue'
 </script>
 
 <template>
