+++ src/components/TalentsView.vue
